+++ src/main/java/com/tech/challenge/tech_challenge/adapters/driver/controllers/OrderController.java
@@ -5,8 +5,10 @@ import com.tech.challenge.tech_challenge.core.application.exceptions.ValidationE
 import com.tech.challenge.tech_challenge.core.application.message.EMessageType;
 import com.tech.challenge.tech_challenge.core.application.message.MessageResponse;
 import com.tech.challenge.tech_challenge.core.domain.entities.Order;
-import com.tech.challenge.tech_challenge.core.domain.services.OrderService;
 import com.tech.challenge.tech_challenge.core.domain.services.extended.OrderClientService;
+import com.tech.challenge.tech_challenge.core.domain.useCases.CreateOrderUseCase;
+import com.tech.challenge.tech_challenge.core.domain.useCases.FindOrderByIdUseCase;
+import com.tech.challenge.tech_challenge.core.domain.useCases.FindOrdersUseCase;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.ArraySchema;
 import io.swagger.v3.oas.annotations.media.Content;
@@ -26,7 +28,13 @@ import java.util.UUID;
 public class OrderController {
 
     @Autowired
-    OrderService orderService;
+    private FindOrdersUseCase findOrdersUseCase;
+
+    @Autowired
+    private FindOrderByIdUseCase findOrderByIdUseCase;
+
+    @Autowired
+    private CreateOrderUseCase createOrderUseCase;
 
     @Autowired
     OrderClientService orderClientService;
@@ -45,7 +53,7 @@ public class OrderController {
             }
     )
     public ResponseEntity all(){
-        return ResponseEntity.status(HttpStatus.OK).body(orderService.list());
+        return ResponseEntity.status(HttpStatus.OK).body(findOrdersUseCase.execute());
     }
 
     @GetMapping("/order/{id}")
@@ -64,7 +72,7 @@ public class OrderController {
     )
     public ResponseEntity one(@PathVariable UUID id) {
         try {
-            return ResponseEntity.status(HttpStatus.OK).body(orderService.getById(id));
+            return ResponseEntity.status(HttpStatus.OK).body(findOrderByIdUseCase.execute(id));
         }catch (ResourceNotFoundException ex){
             return ResponseEntity
                     .status(HttpStatus.NOT_FOUND)
@@ -87,7 +95,7 @@ public class OrderController {
     )
     public ResponseEntity create(@RequestBody Order order) {
         try {
-            return ResponseEntity.status(HttpStatus.CREATED).body(orderService.create(order));
+            return ResponseEntity.status(HttpStatus.CREATED).body(createOrderUseCase.execute(order));
         }catch (ValidationException ex){
             return ResponseEntity
                     .status(HttpStatus.BAD_REQUEST)
