+++ src/main/java/com/tech/challenge/tech_challenge/core/domain/services/extended/OrderClientService.java
@@ -6,15 +6,17 @@ import com.tech.challenge.tech_challenge.core.application.exceptions.ValidationE
 import com.tech.challenge.tech_challenge.core.domain.entities.Client;
 import com.tech.challenge.tech_challenge.core.domain.entities.Order;
 import com.tech.challenge.tech_challenge.core.domain.services.ClientService;
-import com.tech.challenge.tech_challenge.core.domain.services.OrderService;
+import com.tech.challenge.tech_challenge.core.domain.useCases.FindOrderByIdUseCase;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.UUID;
 
 @Service
-public class OrderClientService extends OrderService{
+public class OrderClientService {
 
+    @Autowired
+    private FindOrderByIdUseCase findOrderByIdUseCase;
     @Autowired
     private OrderRepository orderRepository;
 
@@ -22,7 +24,7 @@ public class OrderClientService extends OrderService{
     private ClientService clientService;
 
     public Order addClientToOrder(UUID orderId, UUID clientId) throws ValidationException, ResourceNotFoundException {
-        Order order = getById(orderId);
+        Order order = findOrderByIdUseCase.execute(orderId);
         Client client = clientService.getById(clientId);
         client.validate();
