+++ src/main/java/com/tech/challenge/tech_challenge/core/domain/services/extended/OrderPaymentService.java
@@ -7,16 +7,17 @@ import com.tech.challenge.tech_challenge.core.domain.entities.EOrderStatus;
 import com.tech.challenge.tech_challenge.core.domain.entities.EPaymentStatus;
 import com.tech.challenge.tech_challenge.core.domain.entities.Order;
 import com.tech.challenge.tech_challenge.core.domain.entities.Payment;
-import com.tech.challenge.tech_challenge.core.domain.services.OrderService;
 import com.tech.challenge.tech_challenge.core.domain.services.PaymentService;
 import com.tech.challenge.tech_challenge.core.domain.services.QueueService;
+import com.tech.challenge.tech_challenge.core.domain.useCases.FindOrderByIdUseCase;
+import com.tech.challenge.tech_challenge.core.domain.useCases.UpdateOrderUseCase;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.UUID;
 
 @Service
-public class OrderPaymentService extends OrderService {
+public class OrderPaymentService {
     @Autowired
     private OrderRepository orderRepository;
 
@@ -26,25 +27,31 @@ public class OrderPaymentService extends OrderService {
     @Autowired
     private QueueService queueService;
 
+    @Autowired
+    private FindOrderByIdUseCase findOrderByIdUseCase;
+
+    @Autowired
+    private UpdateOrderUseCase updateOrderUseCase;
+
     public Order addPaymentToOrder(UUID orderId) throws ResourceNotFoundException, ValidationException {
-        Order order = getById(orderId);
+        Order order = findOrderByIdUseCase.execute(orderId);
 
         Payment createdPayment = paymentService.createNewPayment(order.getPrice());
         order.setStatus(EOrderStatus.PAYMENT_PENDING);
         order.setPayment(createdPayment);
 
-        return update(order);
+        return updateOrderUseCase.execute(order);
     }
 
     public Payment updateOrderPayment(UUID orderId, UUID paymentId, EPaymentStatus status) throws ResourceNotFoundException, ValidationException {
         Payment payment = paymentService.getById(paymentId);
         payment.setSatus(status);
 
-        Order order = getById(orderId);
+        Order order = findOrderByIdUseCase.execute(orderId);
         order.setPayment(payment);
 
         Payment updatedPayment = paymentService.update(payment);
-        update(order);
+        updateOrderUseCase.execute(order);
 
         if(status == EPaymentStatus.PAID) {
             queueService.receiveOrder(order);
