+++ src/main/java/com/tech/challenge/tech_challenge/core/domain/useCases/UpdateOrderUseCase.java
@@ -1,0 +1,20 @@
+package com.tech.challenge.tech_challenge.core.domain.useCases;
+
+import com.tech.challenge.tech_challenge.adapters.driven.infra.repositories.OrderRepository;
+import com.tech.challenge.tech_challenge.core.application.exceptions.ValidationException;
+import com.tech.challenge.tech_challenge.core.domain.entities.Order;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+@Service
+public class UpdateOrderUseCase {
+
+    @Autowired
+    private OrderRepository orderRepository;
+
+    public Order execute(Order order) throws ValidationException {
+        order.validate();
+
+        return orderRepository.save(order);
+    }
+}
